Revert "Changed seconds from Int to Double, timer now ticks every 0.1 second"
This reverts commit 3034bdca54ed891bf3a5484939881d2d0a28d721.

+++ EggTimer/Model/Egg.swift
@@ -6,7 +6,7 @@ struct Egg: Identifiable {
     let name: String
     let imageName: String
     let eggType: EggType
-    let cookingTime: Double
+    let cookingTime: Int
 
     init(eggType: EggType) {
         self.eggType = eggType
@@ -15,15 +15,15 @@ struct Egg: Identifiable {
         case .soft:
             name = L10n.soft
             imageName = Asset.softEgg.name
-            cookingTime = 10.0
+            cookingTime = 10
         case .medium:
             name = L10n.medium
             imageName = Asset.mediumEgg.name
-            cookingTime = 20.0
+            cookingTime = 20
         case .hard:
             name = L10n.hard
             imageName = Asset.hardEgg.name
-            cookingTime = 30.0
+            cookingTime = 30
         }
     }
 }
